grain: Make sure the shaders can be loaded in CF

+++ src/glsl/render.frag.glsl
@@ -2,7 +2,7 @@
 #include "archetype/render.glsl"
 
 layout (set = GRAIN_UNIFORM_SET, binding = 0) uniform uniform_block {
-	uint grain__pool_size;
+	int grain__pool_size;
 };
 
 layout(location = 0) flat in uint v_region;
@@ -11,7 +11,7 @@ layout(location = 1) flat in uint v_lid;
 layout(location = 0) out vec4 output;
 
 void main() {
-	uint flat   = v_region * grain__pool_size + v_lid;
+	uint flat   = v_region * uint(grain__pool_size) + v_lid;
 	ivec2 size  = textureSize(grain__texture_0, 0);
 	ivec2 texel = ivec2(int(flat) % size.x, int(flat) / size.x);
 
@@ -19,7 +19,7 @@ void main() {
 	ModuleParams params = grain__load_ModuleParams(v_region);
 	SystemClock clock = grain__load_SystemClock(v_region);
 
-	Schedule sch = schedule(v_lid, grain__pool_size, clock);
+	Schedule sch = schedule(v_lid, uint(grain__pool_size), clock);
 	uint gen = sch.gen + clock.gen_base;
 
 	srand(flat, gen);
