+++ src/glsl/render.vert.glsl
@@ -2,7 +2,7 @@
 #include "archetype/render.glsl"
 
 layout (set = GRAIN_UNIFORM_SET, binding = 0) uniform uniform_block {
-	uint grain__pool_size;
+	int grain__pool_size;  // CF does not support uint uniform
 };
 
 layout(location = 0) flat out uint v_region;
@@ -10,11 +10,11 @@ layout(location = 1) flat out uint v_lid;
 
 void main() {
 	uint inst   = uint(gl_InstanceIndex);
-	uint packed = inst / grain__pool_size; // position in the culled draw list
-	uint lid    = inst % grain__pool_size; // slot within the system's pool
+	uint packed = inst / uint(grain__pool_size); // position in the culled draw list
+	uint lid    = inst % uint(grain__pool_size); // slot within the system's pool
 	uint region = grain__load_draw_region(packed);  // which texture region it occupies
 
-	uint flat   = region * grain__pool_size + lid;
+	uint flat   = region * uint(grain__pool_size) + lid;
 
 	ivec2 size  = textureSize(grain__texture_0, 0);
 	ivec2 texel = ivec2(int(flat) % size.x, int(flat) / size.x);
@@ -23,7 +23,7 @@ void main() {
 	ModuleParams params = grain__load_ModuleParams(region);
 	SystemClock clock = grain__load_SystemClock(region);
 
-	Schedule sch = schedule(lid, grain__pool_size, clock);
+	Schedule sch = schedule(lid, uint(grain__pool_size), clock);
 	uint gen = sch.gen + clock.gen_base;
 
 	srand(flat, gen);
